add sweet alert , show warning before delete any product

+++ src/Pages/MyEquipmentList.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../Provider/AuthProvider";
 import { Link, useLoaderData } from "react-router-dom";
 import Navbar from "../Components/Navbar";
+import Swal from "sweetalert2";
 
 const MyEquipmentList = () => {
     const loadedItems = useLoaderData();
@@ -16,9 +17,36 @@ const MyEquipmentList = () => {
     }, [user, loadedItems]);
 
     const handleDeleteProduct = (id) => {
-        fetch(`http://localhost:5000/product/${id}`,{
-            method: 'DELETE'
-        })
+
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                fetch(`http://localhost:5000/product/${id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        if (data.deletedCount > 0) {
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your Product has been deleted.",
+                                icon: "success"
+                            });
+                            const remaining = items.filter(item => item._id !== id);
+                            setItems(remaining);
+                        }
+                    })
+            }
+        });
     };
 
     return (
